docker workflow: updated tags type priority values (#201)
## Description

In reverse PRs (main -> branch) the docker build uses the tag main
instead of branch, allowing unapproved code to reach main and making the
content of main unpredictable. This fix resolves the issue tanks to the
updated priorities on tags type, allowing the PR and branch name to be
populated and used as tags, instead of the only "main" (that usually had
high priority).

## Changes Made

- Added explicit priorities to the tags type in the "docker.yaml"
workflow file.
- Raised event=pr priority to '800'
- Raised event=branch priority to '800'

## Related Issues

Fixes #185

## Checklist

- [X] I have used a PR title that is descriptive enough for a release
note.
- [X] I have tested these changes locally.
- [ ] I have added appropriate tests or updated existing tests.
- [ ] I have tested these changes on a cluster [name of the cluster] /
customer [name of the customer]
- [X] I have added appropriate documentation or updated existing
documentation.

---------

Co-authored-by: cosimomeli <11272734+cosimomeli@users.noreply.github.com>
Co-authored-by: chicco785 <777218+chicco785@users.noreply.github.com>