workflows: use official action to create github app tokens and support additional repositories scoping (#207)
## Description

This PR modifies workflows to use official action to create github app
tokens and support additional repositories scoping. The additional
scoping is required for actions that import content from other private
repositories that the current one. For example in the golang workflow,
when not using vendoring.

## Changes Made

* use official action to create github app tokens
* support additional repositories scoping

## Related Issues

N/A

## Checklist

- [x] I have used a PR title that is descriptive enough for a release
note.
- [x] I have tested these changes locally.
- [x] I have added appropriate tests or updated existing tests.
- [ ] I have tested these changes on a cluster [name of the cluster] /
customer [name of the customer]
- [ ] I have added appropriate documentation or updated existing
documentation.

---------

Co-authored-by: cosimomeli <11272734+cosimomeli@users.noreply.github.com>